# Via Nick Merwin (32) and Michael Hartl (9)
* polytex-template-59253610: (121 commits)
  Fiddle with the cover
  Add polish
  Improve the CSS for code listings
  Add custom hyphenation
  Finish 1st cut of PolyTeX template
  Add newlines between chapters
  Update the template, including a bug-catcher
  Make the fixtures more robust
  Make anal changes
  Restore underlining in footnote links
  Add some template polish
  Add a bunch of explanatory material
  Add a third template chapter
  Issue pull request for branch more-anal-changes-59602028
  Finish making anal changes
  Finish anal changes to lib/
  Make anal changes
  Issue pull request for branch handle-images-58379736
  Add image and imagebox CSS
  Define image & image w/ box
  ...

+++ lib/polytexnic/builders/pdf.rb
@@ -3,18 +3,28 @@ module Polytexnic
     class Pdf < Builder
       include Polytexnic::Output
 
-      def build!
-        # Build the PolyTeX filename so it accepts both 'foo' and 'foo.tex'.
+      def build!(options={})
         if markdown_directory?
+          # Build the HTML to produce PolyTeX as a side-effect,
+          # then update the manifest to reduce PDF generation
+          # to a previously solved problem.
           Polytexnic::Builders::Html.new.build!
           @manifest = Polytexnic::BookManifest.new(source: :polytex)
         end
+        # Build the PolyTeX filename so it accepts both 'foo' and 'foo.tex'.
         basename = File.basename(@manifest.filename, '.tex')
         book_filename = basename + '.tex'
+
+        # In debug mode, execute `xelatex` and exit.
+        if options[:debug]
+          execute "#{xelatex} #{book_filename}"
+          return    # only gets called in test env
+        end
+
         polytex_filenames = @manifest.chapter_file_paths << book_filename
         polytex_filenames.delete('chapters/frontmatter.tex')
         polytex_filenames.each do |filename|
-          puts filename
+          puts filename unless options[:quiet] || options[:silent]
           polytex = File.open(filename) { |f| f.read }
           latex   = Polytexnic::Core::Pipeline.new(polytex).to_latex
           if filename == book_filename
@@ -27,36 +37,53 @@ module Polytexnic
           end
         end
         write_pygments_file(:latex)
-        write_polytexnic_commands_file
-        cmd = "#{xelatex} #{Polytexnic::Utils.tmpify(book_filename)}"
+        copy_polytexnic_sty
+        build_pdf = "#{xelatex} #{Polytexnic::Utils.tmpify(book_filename)}"
         # Run the command twice to guarantee up-to-date cross-references.
-        system("#{cmd} && #{cmd}")
-        rename_pdf(basename)
+        # Renaming the PDF in the command is necessary because `execute`
+        # below uses `exec` (except in tests, where it breaks). Since `exec`
+        # causes the Ruby process to end, any Ruby code after `exec`
+        # is ignored.
+        # (The reason for using `exec` is so that LaTeX errors get emitted to
+        # the screen rather than just hanging the process.
+        cmd = "#{build_pdf} ; #{build_pdf} ; #{rename_pdf(basename)}"
+        # Here we use `system` when making a preview because the preview command
+        # needs to run after the main PDF build.
+        if options[:quiet] || options[:silent]
+          silence { options[:preview] ? system(cmd) : execute(cmd) }
+        else
+          options[:preview] ? system(cmd) : execute(cmd)
+        end
       end
 
       private
 
+        # Returns the `xelatex` executable.
+        # The `xelatex` program is roughly equivalent to the more standard
+        # `pdflatex`, but has better support for Unicode.
         def xelatex
           filename = `which xelatex`.chomp
           message  = "Install LaTeX (http://latex-project.org/ftp.html)"
           @xelatex ||= executable(filename, message)
         end
 
-        # Renames the temp PDF so that it matches the original filename.
+        # Returns the command to rename the temp PDF.
+        # The purpose is to match the original filename.
         # For example, foo_bar.tex should produce foo_bar.pdf.
         # While we're at it, we move it to the standard ebooks/ directory.
         def rename_pdf(basename)
           tmp_pdf = basename + '.tmp.pdf'
           pdf     = basename + '.pdf'
           mkdir('ebooks')
-          FileUtils.mv(tmp_pdf, File.join('ebooks', pdf))
+          "mv -f #{tmp_pdf} #{File.join('ebooks', pdf)}"
         end
 
-        # Writes out the PolyTeXnic commands from polytexnic-core.
-        def write_polytexnic_commands_file
-          File.open('polytexnic_commands.sty', 'w') do |f|
-            f.write(Polytexnic::Core::Utils.new_commands)
-          end
+        # Copies the PolyTeXnic style file to ensure it's always fresh.
+        def copy_polytexnic_sty
+          polytexnic_sty = 'polytexnic.sty'
+          source_sty     = File.join(File.dirname(__FILE__),
+                                     "../template/#{polytexnic_sty}")
+          FileUtils.cp source_sty, polytexnic_sty
         end
     end
   end
